\name{analisisnumerico-package}
\alias{analisisnumerico-package}
\alias{analisisnumerico}
\docType{package}
\title{Biblioteca de análisis numérico}
\description{Metodos de raices, autovalores, potencia, QR y Householder en R base.}
\details{Use help(package="analisisnumerico") para ver el indice. Cada funcion exportada tiene una pagina propia con argumentos, resultados, ejemplos y limitaciones.}
\seealso{\code{\link{secante}}, \code{\link{grafico_secante}}, \code{\link{regula_falsi}}, \code{\link{biseccion}}, \code{\link{punto_fijo}}, \code{\link{cociente_rayleigh}}, \code{\link{metodo_potencia}}, \code{\link{metodo_potencia_inversa}}, \code{\link{valores_vectores_propios}}, \code{\link{diagonalizar_matriz}}, \code{\link{descomposicion_qr}}, \code{\link{algoritmo_qr}}, \code{\link{reflector_householder}}, \code{\link{householder_reduction}}, \code{\link{multiplicidad_espectral}}, \code{\link{verificar_propiedades_espectrales}}, \code{\link{espacio_propio}}, \code{\link{polinomio_caracteristico}}, \code{\link{householder_qr}}}
